requirements.txt
pytesseract
pillow
gunicorn
moondream